fix: move rbac resources and logic to addons
