+++ pkg/resources/addons/const.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rbac
+package addons
 
 const (
 	KubeSystemNamespace = "kube-system"
