+++ pkg/resources/addons/rbac.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rbac
+package addons
 
 import (
 	corev1 "k8s.io/api/core/v1"
